+++ xerparser/schemas/rsrc.py
@@ -1,6 +1,9 @@
 # xerparser
 # rsrc.py
 
+from typing import Any
+from xerparser.schemas.udftype import UDFTYPE
+
 
 class RSRC:
     """
@@ -26,6 +29,7 @@ class RSRC:
         self.name: str = data["rsrc_name"]
         self.short_name: str = data["rsrc_short_name"]
         self.type: str = data["rsrc_type"]
+        self.user_defined_fields: dict[UDFTYPE, Any] = {}
 
     def __eq__(self, __o: "RSRC") -> bool:
         return all(